[LibOS] Change `mmap` signature so that the file offset is unsigned
On Linux `mmap` syscall takes the file offset as `unsigned long`, only
the POSIX wrapper has it as `off_t` (which is signed).

+++ LibOS/shim/include/shim_fs.h
@@ -41,7 +41,7 @@ struct shim_fs_ops {
 
     /* mmap: mmap handle to address */
     int (*mmap)(struct shim_handle* hdl, void** addr, size_t size, int prot, int flags,
-                off_t offset);
+                size_t offset);
 
     /* flush: flush out user buffer */
     int (*flush)(struct shim_handle* hdl);
